+++ github-runner-manager/tests/unit/test_github_client.py
@@ -20,7 +20,6 @@ from fastcore.net import (  # pylint: disable=no-name-in-module
 )
 from requests import HTTPError as RequestsHTTPError
 
-import github_runner_manager.github_client
 from github_runner_manager.configuration.github import GitHubOrg, GitHubRepo
 from github_runner_manager.github_client import GithubClient, GithubRunnerNotFoundError
 from github_runner_manager.manager.models import InstanceID, RunnerIdentity, RunnerMetadata
@@ -281,7 +280,7 @@ def test_github_api_http_error(github_client: GithubClient, job_stats_raw: JobSt
         )
 
 
-def test_list_runners(github_client: GithubClient, monkeypatch: pytest.MonkeyPatch):
+def test_list_runners(github_client: GithubClient):
     """
     arrange: A mocked Github Client that returns two runners, one for the requested prefix.
     act: Call list_runners with the prefix.
@@ -319,12 +318,10 @@ def test_list_runners(github_client: GithubClient, monkeypatch: pytest.MonkeyPat
         ],
     }
 
-    github_client._client.last_page.return_value = 1
-    github_client._client.actions.list_self_hosted_runners_for_repo.side_effect = response
-
-    pages = MagicMock()
-    pages.return_value = [response]
-    monkeypatch.setattr(github_runner_manager.github_client, "pages", pages)
+    github_client._client.actions.list_self_hosted_runners_for_repo.side_effect = [
+        response,
+        {"runners": []},
+    ]
 
     github_repo = GitHubRepo(owner=secrets.token_hex(16), repo=secrets.token_hex(16))
     runners = github_client.list_runners(path=github_repo, prefix="current-unit-0")
@@ -516,7 +513,7 @@ def test_get_runner_context_org(github_client: GithubClient, monkeypatch: pytest
 
     instance_id = InstanceID.build("test-runner")
 
-    def _mock_generate_runner_jitconfig_for_org(org, name, runner_group_id, labels, timeout):
+    def _mock_generate_runner_jitconfig_for_org(org, name, runner_group_id, labels):
         """Mocked generate_runner_jitconfig_for_org."""
         assert org == "theorg"
         assert name == instance_id.name