+++ terraform/charm/tests/main.tf
@@ -6,7 +6,7 @@ terraform {
   required_providers {
     juju = {
       source  = "juju/juju"
-      version = ">= 0.11.0, < 1.0.0"
+      version = ">= 1.0, < 2.0"
     }
   }
 }
@@ -19,16 +19,16 @@ resource "juju_model" "test_model" {
 
 # Provision a machine and reuse its ID in module placements
 resource "juju_machine" "m0" {
-  model = juju_model.test_model.name
-  base  = "ubuntu@22.04"
-  name  = "machine_0"
+  model_uuid = juju_model.test_model.uuid
+  base       = "ubuntu@22.04"
+  name       = "machine_0"
 }
 
 # Scenario: two applications deployed on a single machine
 module "runner_a" {
   source      = "./.."
   app_name    = "github-runner-a"
-  model       = juju_model.test_model.name
+  model_uuid  = juju_model.test_model.uuid
   channel     = "latest/edge"
   revision    = null
   base        = "ubuntu@22.04"
@@ -41,7 +41,7 @@ module "runner_a" {
 module "runner_b" {
   source      = "./.."
   app_name    = "github-runner-b"
-  model       = juju_model.test_model.name
+  model_uuid  = juju_model.test_model.uuid
   channel     = "latest/edge"
   revision    = null
   base        = "ubuntu@22.04"